Enable usage fault detection (divide by 0 and "unintentional" misaligned accesses)--thread is stopped when exception occurs.  Fixed some "unintentional" misaligned accesses.  Other assembly code improvements.

+++ mos.h
@@ -120,7 +120,7 @@ typedef struct {
 } MosWaitMuxEntry;
 
 typedef struct {
-    u16 num;
+    u32 num;
     MosWaitMuxEntry * entries;
 } MosWaitMux;
